requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
geographiclib==1.52
geopy==2.1.0
idna==2.10
knmi-py==0.1.8
multidict==5.1.0
numpy==1.21.0
pandas==1.3.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
